Change int parsing for float on the time lost param

+++ manager/api/views.py
@@ -1264,7 +1264,7 @@ def jira(request):
                 "customfield_15602": "on"
                 if int(full_request.get("level", 0)) >= 100
                 else "off",  # Is Urgent?
-                "customfield_16702": int(
+                "customfield_16702": float(
                     full_request.get("time_lost", 0)
                 ),  # Obs. time loss
                 "issuetype": {"id": 12302},
